+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fix unneeded php command use in installto.sh and deluser.sh scripts (#9237)
 - Fix TinyMCE localization installation (#9266)
 - Fix bug where trailing non-ascii characters in email addresses could have been removed in recipient input (#9257)
+- Fix IMAP GETMETADATA command with options - RFC5464
 
 ## Release 1.6.5